feat(useFeedbackWidget): Add a more flexible hook for opening feedback forms
