+++ static/app/components/feedback/widget/useFeedbackWidget.tsx
@@ -1,13 +1,7 @@
 import type {RefObject} from 'react';
 import {useEffect} from 'react';
-import * as Sentry from '@sentry/react';
 
-import {t} from 'sentry/locale';
-import ConfigStore from 'sentry/stores/configStore';
-import {useLegacyStore} from 'sentry/stores/useLegacyStore';
-import useAsyncSDKIntegrationStore from 'sentry/views/app/asyncSDKIntegrationProvider';
-
-type FeedbackIntegration = NonNullable<ReturnType<typeof Sentry.getFeedback>>;
+import {useFeedback} from 'sentry/components/feedback/widget/useFeedback';
 
 interface Props {
   buttonRef?: RefObject<HTMLButtonElement> | RefObject<HTMLAnchorElement>;
@@ -20,26 +14,12 @@ export default function useFeedbackWidget({
   formTitle,
   messagePlaceholder,
 }: Props) {
-  const config = useLegacyStore(ConfigStore);
-  const {state} = useAsyncSDKIntegrationStore();
-
-  // TODO(ryan953): remove the fallback `?? Sentry.getFeedback()` after
-  // getsentry is calling `store.add(feedback);`
-  const feedback =
-    (state.Feedback as FeedbackIntegration | undefined) ?? Sentry.getFeedback();
+  const {feedback, options} = useFeedback({formTitle, messagePlaceholder});
 
   useEffect(() => {
     if (!feedback) {
       return undefined;
     }
-
-    const options = {
-      colorScheme: config.theme === 'dark' ? ('dark' as const) : ('light' as const),
-      buttonLabel: t('Give Feedback'),
-      submitButtonLabel: t('Send Feedback'),
-      messagePlaceholder: messagePlaceholder ?? t('What did you expect?'),
-      formTitle: formTitle ?? t('Give Feedback'),
-    };
 
     if (buttonRef) {
       if (buttonRef.current) {
@@ -53,7 +33,7 @@ export default function useFeedbackWidget({
     }
 
     return undefined;
-  }, [buttonRef, config.theme, feedback, formTitle, messagePlaceholder]);
+  }, [buttonRef, feedback, options]);
 
   return feedback;
 }
